Refactor OpType and constants into the types package.
